Redirect to current page, not model list page
If the buttons are used in a view which isn't the list page of the model,
redirecting to the list page is unexpected.

Example: You're in the virtual machine detail view, showing a table from a
plugin. When you click edit/delete on an item in that table, you expect
to be returned to the virtual machine detail view, not the list view of the
model you edited/deleted.

This behaviour is the same as the core ButtonsColumn.